refactor: move `subscribe` code to `execute` file
to preserve history for further integration
