+++ src/execution/index.ts
@@ -1,10 +1,12 @@
 export { pathToArray as responsePathAsArray } from '../jsutils/Path';
 
 export {
+  createSourceEventStream,
   execute,
   executeSync,
   defaultFieldResolver,
   defaultTypeResolver,
+  subscribe,
 } from './execute';
 
 export type {
@@ -12,8 +14,6 @@ export type {
   ExecutionResult,
   FormattedExecutionResult,
 } from './execute';
-
-export { subscribe, createSourceEventStream } from './subscribe';
 
 export {
   getArgumentValues,
